requirements.txt
torch==2.3.0
vllm==0.4.3
flash-attn==2.5.8
transformers==4.44.2
langchain>=0.2.16
langchain-openai>=0.1.25
langchain-community>=0.2.17
llama-index>=0.11.9
llama-index-llms-vllm>=0.2.0
llama-index-llms-langchain>=0.4.1
llama-index-llms-openai>=0.2.5
llama-index-llms-anthropic>=0.3.1
llama-index-embeddings-huggingface>=0.3.1
llama-index-llms-huggingface-api>=0.2.0
llama-index-vector-stores-milvus>=0.2.3
llama-index-postprocessor-flag-embedding-reranker>=0.2.0
flagembedding>=1.2.11
peft>=0.12.0
ray>=2.35.0
pymilvus>=2.4.6
python-dotenv>=1.0.1
ipython>=8.27.0
docling>=1.11.0
docling-core>=1.3.0
numpy<2.0.0
streamlit>=1.38.0
httpx>=0.27.2
fastapi>=0.114.2
pydantic>=2.9.1
uvicorn>=0.30.6